Migration to Vitest completion
