+++ change/@itwin-map-layers-e4e0f3ca-b87b-4fe2-aa2c-57a03c83157d.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "Test Migration to Vitest",
+  "packageName": "@itwin/map-layers",
+  "email": "bipin.pachhai@bentley.com",
+  "dependentChangeType": "none"
+}
