+++ packages/itwin/map-layers/package.json
@@ -19,7 +19,7 @@
     "lint": "eslint -f visualstudio \"./src/**/*.{ts,tsx}\" 1>&2",
     "lintfix": "eslint --fix -f visualstudio \"./src/**/*.{ts,tsx}\" 1>&2",
     "pseudolocalize": "betools pseudolocalize --englishDir=./src/public/locales/en --out=./src/public/locales/en-PSEUDO",
-    "test": "vitest --run src/testme/",
+    "test": "vitest --run",
     "rebuild": "npm run -s clean && npm run -s build"
   },
   "repository": {
