+++ packages/itwin/map-layers/src/test/MapLayerSettingsService.test.ts
@@ -2,8 +2,6 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import * as chai from "chai";
-import * as sinon from "sinon";
 import type { GuidString } from "@itwin/core-bentley";
 import { Guid } from "@itwin/core-bentley";
 import { MapLayerSource, MockRender } from "@itwin/core-frontend";
@@ -11,7 +9,6 @@ import { MapLayerSource, MockRender } from "@itwin/core-frontend";
 import { MapLayerPreferences } from "../MapLayerPreferences";
 import { restore, setup } from "./UserPreferencesMock";
 
-chai.should();
 describe("MapLayerPreferences", () => {
   const iTwinId: GuidString = Guid.createValue();
   const iModelId: GuidString = Guid.createValue();
@@ -21,9 +18,10 @@ describe("MapLayerPreferences", () => {
     await MockRender.App.startup();
     setup();
   });
+
   afterEach(async () => {
     restore();
-    sinon.restore();
+    vi.restoreAllMocks();
     await MockRender.App.shutdown();
   });
 
@@ -34,27 +32,21 @@ describe("MapLayerPreferences", () => {
       formatId: "test12345",
       transparentBackground: true,
     });
-    chai.assert.isDefined(layer);
+    expect(layer).toBeDefined();
     let sources = await MapLayerPreferences.getSources(iTwinId, iModelId);
-    let foundSource = sources.some((value) => {
-      return value.name === testName;
-    });
-    chai.assert.isFalse(foundSource, "expect not to find the source as it has not been stored yet");
+    let foundSource = sources.some((value) => value.name === testName);
+    expect(foundSource).toBe(false);
 
     sources = await MapLayerPreferences.getSources(iTwinId);
-    foundSource = sources.some((value) => {
-      return value.name === testName;
-    });
-    chai.assert.isFalse(foundSource, "expect not to find the source as it has not been stored yet");
+    foundSource = sources.some((value) => value.name === testName);
+    expect(foundSource).toBe(false);
 
     const success = await MapLayerPreferences.storeSource(layer!, iTwinId, iModelId, false);
-    chai.assert.isTrue(success);
+    expect(success).toBe(true);
 
     sources = await MapLayerPreferences.getSources(iTwinId, iModelId);
-    foundSource = sources.some((value) => {
-      return value.name === testName;
-    });
-    chai.assert.isTrue(foundSource);
+    foundSource = sources.some((value) => value.name === testName);
+    expect(foundSource).toBe(true);
   });
 
   it("should store and retrieve layer without ModelId", async () => {
@@ -65,21 +57,17 @@ describe("MapLayerPreferences", () => {
       transparentBackground: true,
     });
 
-    chai.assert.isDefined(layer);
+    expect(layer).toBeDefined();
     let sources = await MapLayerPreferences.getSources(iTwinId);
-    let foundSource = sources.some((value) => {
-      return value.name === testName;
-    });
-    chai.assert.isFalse(foundSource, "expect not to find the source as it has not been stored yet");
+    let foundSource = sources.some((value) => value.name === testName);
+    expect(foundSource).toBe(false);
 
     const success = await MapLayerPreferences.storeSource(layer!, iTwinId);
-    chai.assert.isTrue(success);
+    expect(success).toBe(true);
 
     sources = await MapLayerPreferences.getSources(iTwinId);
-    foundSource = sources.some((value) => {
-      return value.name === testName;
-    });
-    chai.assert.isTrue(foundSource);
+    foundSource = sources.some((value) => value.name === testName);
+    expect(foundSource).toBe(true);
   });
 
   it("should not be able to store model setting if same setting exists as project setting", async () => {
@@ -90,9 +78,9 @@ describe("MapLayerPreferences", () => {
       transparentBackground: true,
     });
     let success = await MapLayerPreferences.storeSource(layer!, iTwinId, iModelId, false);
-    chai.assert.isTrue(success);
+    expect(success).toBe(true);
     success = await MapLayerPreferences.storeSource(layer!, iTwinId, iModelId, true);
-    chai.assert.isFalse(success, "cannot store the iModel setting that conflicts with an iTwin setting");
+    expect(success).toBe(false);
   });
 
   it("should be able to store project setting if same setting exists as project setting", async () => {
@@ -103,9 +91,9 @@ describe("MapLayerPreferences", () => {
       transparentBackground: true,
     });
     let success = await MapLayerPreferences.storeSource(layer!, iTwinId, iModelId, true);
-    chai.assert.isTrue(success);
+    expect(success).toBe(true);
     success = await MapLayerPreferences.storeSource(layer!, iTwinId, iModelId, false);
-    chai.assert.isTrue(success);
+    expect(success).toBe(true);
   });
 
   it("should be able to store the same settings twice without iTwinId and iModelId", async () => {
@@ -116,9 +104,9 @@ describe("MapLayerPreferences", () => {
       transparentBackground: true,
     });
     let success = await MapLayerPreferences.storeSource(layer!, iTwinId);
-    chai.assert.isTrue(success);
+    expect(success).toBe(true);
     success = await MapLayerPreferences.storeSource(layer!, iTwinId);
-    chai.assert.isTrue(success);
+    expect(success).toBe(true);
   });
 
   it("should be able to delete a mapSource stored on project and imodel level", async () => {
@@ -129,15 +117,15 @@ describe("MapLayerPreferences", () => {
       transparentBackground: true,
     });
 
-    chai.assert.isDefined(layer);
+    expect(layer).toBeDefined();
 
-    chai.assert.isTrue(await MapLayerPreferences.storeSource(layer!, iTwinId, iModelId, true));
+    expect(await MapLayerPreferences.storeSource(layer!, iTwinId, iModelId, true)).toBe(true);
     await MapLayerPreferences.deleteByName(layer!, iTwinId, iModelId);
-    chai.assert.isUndefined(await MapLayerPreferences.getByUrl(layer!.url, iTwinId, iModelId));
+    expect(await MapLayerPreferences.getByUrl(layer!.url, iTwinId, iModelId)).toBeUndefined();
 
-    chai.assert.isTrue(await MapLayerPreferences.storeSource(layer!, iTwinId, iModelId, true));
+    expect(await MapLayerPreferences.storeSource(layer!, iTwinId, iModelId, true)).toBe(true);
     await MapLayerPreferences.deleteByName(layer!, iTwinId, iModelId);
-    chai.assert.isUndefined(await MapLayerPreferences.getByUrl(layer!.url, iTwinId, iModelId));
+    expect(await MapLayerPreferences.getByUrl(layer!.url, iTwinId, iModelId)).toBeUndefined();
   });
 
   it("should be able to delete a mapSource stored without iTwinId and iModelId", async () => {
@@ -148,10 +136,10 @@ describe("MapLayerPreferences", () => {
       transparentBackground: true,
     });
 
-    chai.assert.isDefined(layer);
+    expect(layer).toBeDefined();
 
-    chai.assert.isTrue(await MapLayerPreferences.storeSource(layer!, iTwinId));
+    expect(await MapLayerPreferences.storeSource(layer!, iTwinId)).toBe(true);
     await MapLayerPreferences.deleteByName(layer!, iTwinId);
-    chai.assert.isUndefined(await MapLayerPreferences.getByUrl(layer!.url, iTwinId));
+    expect(await MapLayerPreferences.getByUrl(layer!.url, iTwinId)).toBeUndefined();
   });
 });
