+++ packages/itwin/map-layers/src/test/TestUtils.ts
@@ -2,7 +2,6 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-
 import { MapLayersUI } from "../map-layers";
 
 /** @internal */
