+++ packages/itwin/map-layers/src/test/setup.ts
@@ -2,7 +2,6 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-
 beforeEach(() => {
   vi.stubGlobal("fetch", async () => Promise.resolve(new Response()));
 });
